Configure summary omits "S-Lang" when used as screen library.
When using slang, 'configure' prints blank in the "Screen library" name:

    ...
    File system:                Midnight Commander Virtual Filesystem
                                cpio, extfs, fish, ftp, sfs, sftp, smb, tar
    Screen library:
    Mouse support:              gpm and xterm
    X11 events support:         yes
    ...

This occurs when slang is found using pkg-config (as opposed to explicit path
search). The patch fixes this by moving the 'screen_msg' assignment one level
up, to the controlling function (mc_WITH_SLANG) so that it's effective for
however method the library was found.

The problem was mentioned on our mailing list:

    https://mail.gnome.org/archives/mc-devel/2016-March/msg00021.html

Signed-off-by: Andrew Borodin <aborodin@vmail.ru>